+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Twigpack Changelog
 
+## 1.1.0 - 2018-10-09
+### Added
+* Strings passed in to `manifestPath` can now be Yii2 aliases as well
+* Added `craft.twigpack.includeFile()`
+* Added `craft.twigpack.includeFileFromManifest()`
+* Added `craft.twigpack.includeInlineCssTags()`
+* Added `craft.twigpack.includeCriticalCssTags()`
+
 ## 1.0.5 - 2018-09-28
 ### Changed
 * Check via `empty()` rather than `!== null` when checking the manifest for module entries